+++ src/events.c
@@ -19,6 +19,7 @@ static void event_watcher(uint32_t event, void* data, size_t data_length, void* 
 }
 
 static void window_spawn_handler(uint32_t event, void* data, size_t data_length, void* context) {
+  struct table* windows = &g_windows;
   int cid = (intptr_t)context;
   uint32_t wid = *(uint32_t *)(data + 0x8);
   uint64_t sid = *(uint64_t *)data;
@@ -46,8 +47,8 @@ static void window_spawn_handler(uint32_t event, void* data, size_t data_length,
       if (iterator && SLSWindowIteratorGetCount(iterator) > 0) {
         if (SLSWindowIteratorAdvance(iterator)) {
           ITERATOR_WINDOW_SUITABLE(iterator, {
-            border_create(wid, sid);
-            update_window_notifications();
+            windows_window_create(windows, wid, sid);
+            windows_update_notifications(windows);
           });
         }
       }
@@ -57,49 +58,44 @@ static void window_spawn_handler(uint32_t event, void* data, size_t data_length,
     CFRelease(target_ref);
   } else if (event == EVENT_WINDOW_DESTROY) {
     printf("Window Destroyed: %d\n", wid);
-    if (windows_remove_window(&g_windows, wid, sid)) {
-      update_window_notifications();
+    if (windows_window_destroy(windows, wid, sid)) {
+      windows_update_notifications(windows);
     }
 
-    borders_window_focus(get_front_window());
+    windows_window_focus(windows, get_front_window());
   }
 }
 
 static void window_modify_handler(uint32_t event, void* data, size_t data_length, void* context) {
   int cid = (intptr_t)context;
   uint32_t wid = *(uint32_t *)(data);
+  struct table* windows = &g_windows;
 
   if (event == EVENT_WINDOW_MOVE) {
     printf("Window Move: %d\n", wid);
-    borders_move_border(wid);
+    windows_window_move(windows, wid);
   } else if (event == EVENT_WINDOW_RESIZE) {
-    borders_update_border(wid);
+    windows_window_update(windows, wid);
   } else if (event == EVENT_WINDOW_REORDER) {
     printf("Window Reorder: %d\n", wid);
 
     // The update of the front window might not have taken place yet...
     usleep(10000);
 
-    borders_window_focus(get_front_window());
+    windows_window_focus(windows, get_front_window());
   } else if (event == EVENT_WINDOW_LEVEL) {
     printf("Window Level: %d\n", wid);
-    borders_update_border(wid);
+    windows_window_update(windows, wid);
   } else if (event == EVENT_WINDOW_TITLE || event == EVENT_WINDOW_UPDATE) {
     wid = get_front_window();
     printf("Window Focus: %d\n", wid);
-    borders_window_focus(wid);
+    windows_window_focus(windows, wid);
   } else if (event == EVENT_WINDOW_UNHIDE) {
     printf("Window Unhide: %d\n", wid);
-    struct border *border = table_find(&g_windows, &wid);
-    if (border) {
-      SLSOrderWindow(cid, border->wid, -1, border->target_wid);
-    }
+    windows_window_unhide(windows, wid);
   } else if (event == EVENT_WINDOW_HIDE) {
     printf("Window Hide: %d\n", wid);
-    struct border *border = table_find(&g_windows, &wid);
-    if (border) {
-      SLSOrderWindow(cid, border->wid, 0, border->target_wid);
-    }
+    windows_window_hide(windows, wid);
   }
 }
